add initRoutes
