--- core/src/core/route.ts
@@ -1,5 +1,0 @@
-import store from '../../store';
-
-export async function initRoutes() {
-  const pageModel = store.db.model('page');
-}
