+++ core/types/route.ts
@@ -1,4 +1,3 @@
 export type StoreRoute = {
-  isDynamic?: boolean;
   path: string;
 };
